testing float16

+++ kandinsky/kandinsky_train.py
@@ -43,7 +43,7 @@ def log_memory_usage(description, current_memory_usage):
 
     return total_memory
 
-weight_dtype = torch.float32
+weight_dtype = torch.float16
 
 device = torch.device(0)
